openai_api/pyproject.toml
[project]
name = "openai-cli"
version = "0.1.0"
description = "Simple CLI tool for OpenAI ChatGPT API"
authors = [{name = "CLI Tool", email = "tool@example.com"}]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"

[project.scripts]
openai-cli = "openai_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []